pyproject.toml
[tool.poetry]
name = "cloud-octo-apigateway"
version = "0.1.0"
description = ""
authors = ["pazsaragi <pazsaragi40@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.21.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
gunicorn = "^20.1.0"
fastapi = "^0.70.1"
boto3 = "^1.20.24"
pynamodb = "^5.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
black = {version = "^21.12b0", allow-prereleases = true}
awscli-local = "^0.17"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
